去掉无用的MainWindow.xaml 将ProgamInit提出去
